run-ci: fix armerge symbol keeping regex
If we expose all symbols that contain tanker instead of just symbols
that start with tanker, we end up exposing a lot of template symbols
that contain tanker inside.

We link tanker twice in sdk-ios tests because the tests need access to
tankeradmin. Because of that, some symbols of one instance may be
overwritten by the other. This poses a problem because libc++abi is
linked statically and hidden. It contains in particuar a thread-local
variable counting the uncaught exceptions (in addition to other state
stuff). Then an exception thrown by one instance of the library will
increment the counter, and it may be caught by another instance, which
will decrement the second counter, leading to all kinds of undefined
behavior (like std::uncaught_exceptions() returning -1).