fix: Unset deployment target for all deps in Podfile
Some of our libs set low deployment targets, like iOS 8.0
This is below the minimum that XCode supports,
and we don't want random libs to set the target anyways,
we want the target to be inherited from our parent project.

Deleting the property in a post-install script seems to be the
standard Cocoapod solution.
(Rumors suggest this used to happen automatically, but not since 1.0)