feat: Introduce TKRStreamBase, similar to POSBlobInputStream
TKRStreamBase fills a similar role as POSBlobInputStream,
in that it can be used to relatively easily create NSInputStream
subclasses, but doesn't use the concept of data source
and doesn't handle all the different cases that POS supports.

Instead, a subclass of TKRStreamBase can override the read function
to creates a valid NSInputStream implementation.