feat: Convert VerificationOptions to Swift
This is a breaking API change, but necessary
to make it usable from Swift.
Rigth now the `+ options()` objc static constructor is automatically
assumed to be the `init` method in Swift.

So in Swift we _cannot_ call options(), it always just calls init().
EXCEPT there is _also_ an init() in ObjC, and it's NOT the same
as options(), so Swift ends up calling the wrong constructor.

In practice it does the same, since VerificationOptions happens
to have everything false by default, and default init is a memzero.
But this is a VERY subtle broken API that could cause trouble later.